GUI prototyping
This commit contains an eye candy GUI prototype using Avalonia.
A lite theme will also be developed for systems with low resources, and the the eye candy GUI will be optimized to stop rendering when its completely obstructed from view or certain features like the acrylic blur will disable once user unfocuses the window. A replica of the mouse pointer visualiztion feature from Windows 11 is also to be expected in future commits.

I have been considering it, and it might be a better choice to use purely dotnet code instead of golang and dotnet.
Future commits that get rid of the golang parts may be expected if I go through with this.